Slowly transition from Vec2f to Vec2<template>
Button, Sprite's variables. And GraphicSystem's camOffsetRotated variable

+++ Cpp_OpenGL_Fun/src/Sprite.cpp
@@ -55,7 +55,7 @@ namespace OpenGLFun {
 		auto findResult = std::find(texturesList.begin(), texturesList.end(), mTextureFilepath);
 
 		if (findResult != texturesList.end())
-			_selectedTexture = findResult - texturesList.begin();
+			_selectedTexture = static_cast<int>(findResult - texturesList.begin());
 
 		if (ImGui::BeginCombo("##texture list", texturesList[_selectedTexture].c_str())) {
 			for (int i = 0; i < texturesList.size(); i++) {
